New odin version, now at 600 lines
Converted the cons cell storage from a tagged union to
storing the tag in the low bits of the car pointer of
the cell, so that a cell is now 16 bytes (two pointers).

Also split numbers and symbols into separate tags and
parsing numbers early, so that there is no string
conversion in the math routines.

Storing symbols as odin strings in an intern table, but
the reference as a split length and pointer value, and
storing the length in the upper bits of the car pointer
(string pointer in cdr).